+++ src/main/java/com/example/demo/entities/FollowID.java
@@ -3,14 +3,14 @@ package com.example.demo.entities;
 import java.io.Serializable;
 
 public class FollowID implements Serializable {
-    private long follower;
-    private long followed;
+    private Integer follower;
+    private Integer followed;
 
     public FollowID() {
 
     }
 
-    public FollowID(long follower, long followed) {
+    public FollowID(Integer follower, Integer followed) {
         this.follower = follower;
         this.followed = followed;
     }
